+++ request/container.go
@@ -14,7 +14,7 @@ type ChimithequeContextKey string
 type Container struct {
 	PersonEmail    string `json:"PersonEmail"`
 	PersonLanguage string `json:"PersonLanguage"`
-	PersonID       int    `json:"PersonID"`
+	PersonID       int64  `json:"PersonID"`
 	AppURL         string `json:"AppURL"`
 	AppPath        string `json:"AppPath"`
 	BuildID        string `json:"BuildID"`
